chatbot-opensource/requirements.txt
langchain_core
langchain_community
python-dotenv
streamlit